Refactored DBC objects into separate files, all tests passing
