+++ src/dbc.rs
@@ -1,0 +1,387 @@
+#[cfg(feature = "with-serde")]
+extern crate serde;
+#[cfg(feature = "with-serde")]
+#[macro_use]
+extern crate serde_derive;
+
+use std::convert::TryFrom;
+
+use derive_getters::Getters;
+
+use crate::DBCString;
+use crate::Error;
+use crate::message::*;
+use crate::signal::*;
+use crate::nodes::*;
+use crate::env_variables::*;
+use crate::attributes::*;
+
+use crate::parser;
+
+/// Baudrate of network in kbit/s
+#[derive(Copy, Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Serialize))]
+pub struct Baudrate(pub (crate) u64);
+
+impl DBCString for Baudrate {
+    fn dbc_string(&self) -> String {
+        return self.0.to_string()
+    }
+}
+
+/// Version generated by DB editor
+#[derive(Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Serialize))]
+pub struct Version(pub String);
+
+impl DBCString for Version {
+    fn dbc_string(&self) -> String {
+        return format!("VERSION {}", self.0)
+    }
+}
+
+#[derive(Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Serialize))]
+pub struct Symbol(pub String);
+
+impl DBCString for Symbol {
+    fn dbc_string(&self) -> String {
+        return self.0.to_string()
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Getters)]
+#[cfg_attr(feature = "serde", derive(Serialize))]
+pub struct DBC {
+    /// Version generated by DB editor
+    pub (crate) version: Version,
+    pub (crate) new_symbols: Vec<Symbol>,
+    /// Baud rate of network
+    pub (crate) bit_timing: Option<Vec<Baudrate>>,
+    /// CAN network nodes
+    pub (crate) nodes: Vec<Node>,
+    /// Global value table
+    pub (crate) value_tables: Vec<ValueTable>,
+    /// CAN message (frame) details including signal details
+    pub (crate) messages: Vec<Message>,
+    pub (crate) message_transmitters: Vec<MessageTransmitter>,
+    pub (crate) environment_variables: Vec<EnvironmentVariable>,
+    pub (crate) environment_variable_data: Vec<EnvironmentVariableData>,
+    pub (crate) signal_types: Vec<SignalType>,
+    /// Object comments
+    pub (crate) comments: Vec<Comment>,
+    pub (crate) attribute_definitions: Vec<AttributeDefinition>,
+    // undefined
+    // sigtype_attr_list: SigtypeAttrList,
+    pub (crate) attribute_defaults: Vec<AttributeDefault>,
+    pub (crate) attribute_values: Vec<AttributeValueForObject>,
+    /// Encoding for signal raw values
+    pub (crate) value_descriptions: Vec<ValueDescription>,
+    // obsolete + undefined
+    // category_definitions: Vec<CategoryDefinition>,
+    // obsolete + undefined
+    //categories: Vec<Category>,
+    // obsolete + undefined
+    //filter: Vec<Filter>,
+    pub (crate) signal_type_refs: Vec<SignalTypeRef>,
+    /// Signal groups define a group of signals within a message
+    pub (crate) signal_groups: Vec<SignalGroups>,
+    pub (crate) signal_extended_value_type_list: Vec<SignalExtendedValueTypeList>,
+    /// Extended multiplex attributes
+    pub (crate) extended_multiplex: Vec<ExtendedMultiplex>,
+}
+
+#[allow(dead_code)]
+fn dbc_vec_to_string<T: DBCString>(dbc_objects: &Vec<T>, delimiter: &str) -> String {
+    dbc_objects
+    .into_iter()
+    .map(|sym| sym.dbc_string())
+    .collect::<Vec<String>>()
+    .join(delimiter)
+}
+
+impl DBCString for DBC {
+    fn dbc_string(&self) -> String {
+        let mut file_str = String::new();
+        // Version
+        file_str += &self.version.dbc_string();
+
+        // Symbols
+        file_str += "NS_ :\n";
+        file_str += &dbc_vec_to_string::<Symbol>(&self.new_symbols, "\n    ");
+        
+
+        // Baudrates
+        match &self.bit_timing {
+            Some(bauds) => {
+                // confirm delineator
+                file_str += &dbc_vec_to_string::<Baudrate>(&bauds, "\n");
+            }
+            None => {}
+        }
+
+        // Nodes
+        file_str += &dbc_vec_to_string::<Node>(&self.nodes, " ");
+
+        // Value Tables
+        file_str += &dbc_vec_to_string::<ValueTable>(&self.value_tables, " ");
+
+        // Messages
+        file_str += &dbc_vec_to_string::<Message>(&self.messages, "\n");
+        file_str += &dbc_vec_to_string::<MessageTransmitter>(&self.message_transmitters, " ");
+        
+        // Environment Variables
+        file_str += &dbc_vec_to_string::<EnvironmentVariable>(&self.environment_variables, "\n");
+        file_str += &dbc_vec_to_string::<EnvironmentVariableData>(&self.environment_variable_data, "\n");
+
+        // Signal Types
+        file_str += &dbc_vec_to_string::<SignalType>(&self.signal_types, " ");
+
+        // Comments
+        file_str += &dbc_vec_to_string::<Comment>(&self.comments, "\n");
+
+        // Attributes
+        file_str += &dbc_vec_to_string::<AttributeDefinition>(&self.attribute_definitions, "\n");
+        file_str += &dbc_vec_to_string::<AttributeValueForObject>(&self.attribute_values, "\n");
+        file_str += &dbc_vec_to_string::<AttributeDefault>(&self.attribute_defaults, "\n");
+
+        // Value Descriptions
+        file_str += &dbc_vec_to_string::<ValueDescription>(&self.value_descriptions, "\n");
+
+        // Signal Attributes
+        file_str += &dbc_vec_to_string::<SignalTypeRef>(&self.signal_type_refs, "\n");
+        file_str += &dbc_vec_to_string::<SignalGroups>(&self.signal_groups, "\n");
+        file_str += &dbc_vec_to_string::<SignalExtendedValueTypeList>(&self.signal_extended_value_type_list, "\n");
+
+        // Multiplex
+        file_str += &dbc_vec_to_string::<ExtendedMultiplex>(&self.extended_multiplex, "\n");
+
+        return file_str
+    }
+}
+
+impl DBC {
+    /// Read a DBC from a buffer
+    #[allow(clippy::result_large_err)]
+    pub fn from_slice(buffer: &[u8]) -> Result<DBC, Error> {
+        let dbc_in = std::str::from_utf8(buffer).unwrap();
+        Self::try_from(dbc_in)
+    }
+
+    #[allow(clippy::should_implement_trait)]
+    #[deprecated(since = "4.0.0", note = "please use `DBC::try_from` instead")]
+    #[allow(clippy::result_large_err)]
+    pub fn from_str(dbc_in: &str) -> Result<DBC, Error> {
+        let (remaining, dbc) = parser::dbc(dbc_in).map_err(Error::Nom)?;
+        if !remaining.is_empty() {
+            return Err(Error::Incomplete(dbc, remaining));
+        }
+        Ok(dbc)
+    }
+
+    pub fn signal_by_name(&self, message_id: MessageId, signal_name: &str) -> Option<&Signal> {
+        let message = self
+            .messages
+            .iter()
+            .find(|message| message.message_id == message_id);
+
+        if let Some(message) = message {
+            return message
+                .signals
+                .iter()
+                .find(|signal| signal.name == *signal_name);
+        }
+        None
+    }
+
+    /// Lookup a message comment
+    pub fn message_comment(&self, message_id: MessageId) -> Option<&str> {
+        self.comments
+            .iter()
+            .filter_map(|x| match x {
+                Comment::Message {
+                    message_id: ref x_message_id,
+                    ref comment,
+                } => {
+                    if *x_message_id == message_id {
+                        Some(comment.as_str())
+                    } else {
+                        None
+                    }
+                }
+                _ => None,
+            })
+            .next()
+    }
+
+    /// Lookup a signal comment
+    pub fn signal_comment(&self, message_id: MessageId, signal_name: &str) -> Option<&str> {
+        self.comments
+            .iter()
+            .filter_map(|x| match x {
+                Comment::Signal {
+                    message_id: ref x_message_id,
+                    signal_name: ref x_signal_name,
+                    comment,
+                } => {
+                    if *x_message_id == message_id && x_signal_name == signal_name {
+                        Some(comment.as_str())
+                    } else {
+                        None
+                    }
+                }
+                _ => None,
+            })
+            .next()
+    }
+
+    /// Lookup value descriptions for signal
+    pub fn value_descriptions_for_signal(
+        &self,
+        message_id: MessageId,
+        signal_name: &str,
+    ) -> Option<&[ValDescription]> {
+        self.value_descriptions
+            .iter()
+            .filter_map(|x| match x {
+                ValueDescription::Signal {
+                    message_id: ref x_message_id,
+                    signal_name: ref x_signal_name,
+                    ref value_descriptions,
+                } => {
+                    if *x_message_id == message_id && x_signal_name == signal_name {
+                        Some(value_descriptions.as_slice())
+                    } else {
+                        None
+                    }
+                }
+                _ => None,
+            })
+            .next()
+    }
+
+    /// Lookup the extended value for a given signal
+    pub fn extended_value_type_for_signal(
+        &self,
+        message_id: MessageId,
+        signal_name: &str,
+    ) -> Option<&SignalExtendedValueType> {
+        self.signal_extended_value_type_list
+            .iter()
+            .filter_map(|x| {
+                let SignalExtendedValueTypeList {
+                    message_id: ref x_message_id,
+                    signal_name: ref x_signal_name,
+                    ref signal_extended_value_type,
+                } = x;
+                if *x_message_id == message_id && x_signal_name == signal_name {
+                    Some(signal_extended_value_type)
+                } else {
+                    None
+                }
+            })
+            .next()
+    }
+
+    /// Lookup the message multiplexor switch signal for a given message
+    /// This does not work for extended multiplexed messages, if multiple multiplexors are defined for a message a Error is returned.
+    #[allow(clippy::result_large_err)]
+    pub fn message_multiplexor_switch(
+        &self,
+        message_id: MessageId,
+    ) -> Result<Option<&Signal>, Error> {
+        let message = self
+            .messages
+            .iter()
+            .find(|message| message.message_id == message_id);
+
+        if let Some(message) = message {
+            if self
+                .extended_multiplex
+                .iter()
+                .any(|ext_mp| ext_mp.message_id == message_id)
+            {
+                Err(Error::MultipleMultiplexors)
+            } else {
+                Ok(message
+                    .signals
+                    .iter()
+                    .find(|signal| signal.multiplexer_indicator == MultiplexIndicator::Multiplexor))
+            }
+        } else {
+            Ok(None)
+        }
+    }
+}
+
+impl<'a> TryFrom<&'a str> for DBC {
+    type Error = Error<'a>;
+
+    fn try_from(dbc_in: &'a str) -> Result<Self, Self::Error> {
+        let (remaining, dbc) = parser::dbc(dbc_in).map_err(Error::Nom)?;
+        if !remaining.is_empty() {
+            return Err(Error::Incomplete(dbc, remaining));
+        }
+        Ok(dbc)
+    }
+}
+
+/// Object comments
+#[derive(Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Serialize))]
+pub enum Comment {
+    Node {
+        node_name: String,
+        comment: String,
+    },
+    Message {
+        message_id: MessageId,
+        comment: String,
+    },
+    Signal {
+        message_id: MessageId,
+        signal_name: String,
+        comment: String,
+    },
+    EnvVar {
+        env_var_name: String,
+        comment: String,
+    },
+    Plain {
+        comment: String,
+    },
+}
+
+impl DBCString for Comment {
+    fn dbc_string(&self) -> String {
+        return match self {
+            Self::Node{node_name, comment} => {
+                format!("BU_ {} \"{}\"",
+                    node_name,
+                    comment
+                )
+            },
+            Self::Message{message_id, comment} => {
+                format!("BO_ {} \"{}\"",
+                    message_id.dbc_string(),
+                    comment,
+                )
+            },
+            Self::Signal{message_id, signal_name, comment} => {
+                format!("SG_ {} {} \"{}\"",
+                    message_id.dbc_string(),
+                    signal_name,
+                    comment
+                )
+            },
+            Self::EnvVar{env_var_name, comment} => {
+                format!("EV_ {} \"{}\"",
+                    env_var_name,
+                    comment,
+                )
+            },
+            Self::Plain{comment} => format!("\"{}\"", comment),
+        }
+    }
+}
+
